fix: Support cases where request body is a Buffer
When using `serverless-http` to run the api-proxy in AWS Lambda, the
request body is held in a Buffer. This was causing a bad request because
the Buffer needs to be passed along, instead of turned in to a JSON
string.

The call to `proxyReq.write` uses Node's `stream.Writeable` class:
https://nodejs.org/api/stream.html#stream_writable_write_chunk_encoding_callback
supports writing a Buffer as a chunk.