feature: Rewrite the domain of proxied `set-cookie` headers
When running a Lambda function behind API gateway behind CloudFront, the
Host of API gateway ends up getting set as the domain on some cookies,
when it should be the CloudFront domain.

This commit inspects the `set-cookie` headers and replaces the domain
with the configured `cookieDomainRewrite` if the cookie's domain matches
the current host.